Comment out issue-causing code
